perf: allocate less when creating a new instance of Player
this is also a memory reduction trick; you're storing a (cloned)
SocketRef and then a stringified SocketRef.id. This is bad because the
SocketRef.id field exists twice in memory, and not to mention that one
is bigger because it is a String type.
see https://github.com/Stars-Beyond/Horizon-Community-Edition/issues/61